requirements.txt
click==7.1.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.20
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3